# docs/api_specs.yaml
openapi: 3.0.3
info:
    title: Echo API
    version: 1.0.0
    description: |
        Local-first REST API for Echo - a multimodal semantic second brain.
        All processing runs on the user's device. No external network calls are
        made during normal operation. The API never generates content; it only
        stores, retrieves and surfaces the user's own existing notes.
servers:
    - url: http://127.0.0.1:8000/
      description: Local development server

#  REUSABLE COMPONENTS
components:
    schemas:
...
        NoteCreate:
            type: object
            required: [note_type]
            properties:
                note_type:
                    $ref: "#/components/schemas/NoteType"
                title:
                    type: string
                    nullable: true
                    maxLength: 500
                body:
                    type: string
                    nullable: true
                    maxLength: 100000
                show_generated_content:
                    type: boolean
                    default: true
...
        Error500:
            type: object
            required: [error, detail]
            properties:
                error:
                    type: string
                    example: "internal_error"
                detail:
                    type: string
                    example: "An unexpected error occurred."

    #  PARAMETERS
    parameters:
        NoteId:
            name: id
            in: path
            required: true
            schema:
                type: string
                format: uuid
            description: The UUID of the note.
...
        NotesLimitParam:
            name: limit
            in: query
            required: false
            schema:
                type: integer
                minimum: 1
                maximum: 200
                default: 50
            description: >
                Maximum number of notes to return per page. Default of 50
                is intended for general note browsing, not similarity ranking.
        IncludeDeleted:
            name: include_deleted
            in: query
            required: false
            schema:
                type: boolean
                default: false
            description: When true, includes soft-deleted notes in results.
        NoteTypeFilter:
            name: note_type
            in: query
            required: false
            schema:
                $ref: "#/components/schemas/NoteType"
            description: Filter notes by type.
        TagIdFilter:
            name: tag_id
            in: query
            required: false
            schema:
                type: string
                format: uuid
            description: >
                Filter notes to only those assigned this tag. Served by the
                idx_note_tags_tag_id index (reverse lookup from tag to notes).

#  PATHS
paths:
...
    /notes:
        get:
            operationId: listNotes
            summary: List all notes
            description: >
                Returns all non-deleted notes ordered by updated_at descending.
                Attachments are not included by default to keep response size small;
                use GET /notes/{id} for full detail.
            tags: [Notes]
            parameters:
                - $ref: "#/components/parameters/NotesLimitParam"
                - $ref: "#/components/parameters/IncludeDeleted"
                - $ref: "#/components/parameters/NoteTypeFilter"
                - $ref: "#/components/parameters/TagIdFilter"
                - name: search
                  in: query
                  required: false
                  schema:
                      type: string
                      maxLength: 200
                  description: Full-text search against title and body fields.
            responses:
                "200":
                    description: Array of notes.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Note"
                            example:
                                - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                                  note_type: text
                                  title: "Random thought: Why clouds?"
                                  body: "There is so much wonder in clouds. We have always been taught that they are bodies of water vapour, but what if they're so much more?"
                                  embedding_text: "Random thought: Why clouds? There is so much wonder in clouds. We have always been taught that they are bodies of water vapour, but what if they're so much more?"
                                  embedding_status: complete
                                  show_generated_content: true
                                  is_favourite: false
                                  is_deleted: false
                                  created_at: "2025-11-03T14:22:00Z"
                                  updated_at: "2025-11-03T14:30:00Z"
                                  deleted_at: null
                                  word_count: 12
                                  attachments: []
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"

        post:
            operationId: createNote
            summary: Create a new note
            description: >
                Creates a new note and immediately returns it with embedding_status
                set to "pending". All ML processing (caption, OCR, transcription,
                embedding) is dispatched to the background job queue and does not
                block this response. This guarantees note save time under 200ms (!NF-01).
            tags: [Notes]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/NoteCreate"
                        example:
                            note_type: text
                            title: "Random thought: Why clouds?"
                            body: "There is so much wonder in clouds. We have always been taught that they are bodies of water vapour, but what if they're so much more?"
            responses:
                "201":
                    description: Note created successfully.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Note"
                            example:
                                id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                                note_type: text
                                title: "Random thought: Why clouds?"
                                body: "There is so much wonder in clouds. We have always been taught that they are bodies of water vapour, but what if they're so much more?"
                                embedding_text: null
                                embedding_status: pending
                                show_generated_content: true
                                is_favourite: false
                                is_deleted: false
                                created_at: "2025-11-03T14:22:00Z"
                                updated_at: "2025-11-03T14:22:00Z"
                                deleted_at: null
                                word_count: 14
                                attachments: []
                "400":
                    description: Validation error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error400"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
    /notes/{id}:
        get:
            operationId: getNote
            summary: Get a single note
            description: Returns full note detail including all attachments.
            tags: [Notes]
            parameters:
                - $ref: "#/components/parameters/NoteId"
            responses:
                "200":
                    description: Note found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Note"
                "404":
                    description: Note not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"

        patch:
            operationId: updateNote
            summary: Update a note
            description: >
                Partial update. Only supplied fields are modified. If body or title
                changes, a new "embed" job is queued and embedding_status reverts
                to "pending".
            tags: [Notes]
            parameters:
                - $ref: "#/components/parameters/NoteId"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/NoteUpdate"
                        example:
                            body: "There is so much wonder in clouds. We have always been taught that they are bodies of water vapour, but what if they're so much more?"
                            is_favourite: true
            responses:
                "200":
                    description: Note updated.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Note"
                "400":
                    description: Validation error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error400"
                "404":
                    description: Note not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"

        delete:
            operationId: deleteNote
            summary: Soft-delete a note (move to trash)
            description: >
                Sets is_deleted=true and deleted_at to the current timestamp.
                Discards any pending_approval and rejected links for this
                note. Confirmed links and note_tags are preserved but
                excluded from query results while the note is_deleted.
                Queued jobs for this note are cancelled; already-running jobs
                are allowed to finish but their results are discarded rather
                than committed. Updates ChromaDB metadata to is_deleted=true
                so the note is excluded from all future similarity queries.
                The note can be restored via POST /notes/{id}/restore.
                To permanently delete, call DELETE /notes/{id}/permanent.
            tags: [Notes]
            parameters:
                - $ref: "#/components/parameters/NoteId"
            responses:
                "204":
                    description: Note moved to trash. No response body.
                "404":
                    description: Note not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
    /notes/{id}/permanent:
        delete:
            operationId: permanentlyDeleteNote
            summary: Permanently delete a note
            description: >
                Irreversibly removes the note from SQLite, its embedding vector
                from ChromaDB, all associated links, jobs, note_tags, canvas_items
                rows and any files on disk. SQLite row is deleted last so that
                partial failures are recoverable. Note must already be soft-deleted
                (is_deleted=true) before this endpoint can be called.
            tags: [Notes]
            parameters:
                - $ref: "#/components/parameters/NoteId"
            responses:
                "204":
                    description: Note permanently deleted. No response body.
                "400":
                    description: Note is not in trash.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error400"
                "404":
                    description: Note not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
    /notes/{id}/restore:
        post:
            operationId: restoreNote
            summary: Restore a soft-deleted note
            description: >
                Sets is_deleted=false and deleted_at=null. Restores visibility
                of any previously confirmed links and note_tags (these were
                never deleted, only excluded from queries while is_deleted was
                true). Queues a fresh "generate_links" job, since any
                pending_approval links that existed before deletion were
                discarded rather than preserved. Updates ChromaDB metadata to
                is_deleted=false so the note becomes eligible again as a match
                in other notes' similarity queries.
            tags: [Notes]
            parameters:
                - $ref: "#/components/parameters/NoteId"
            responses:
                "200":
                    description: Note restored.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Note"
                "400":
                    description: Note is not currently deleted.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error400"
                "404":
                    description: Note not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
...